+++ rat-theme/changes.md
@@ -1,6 +1,7 @@
 # 0.29.0
 
 * upgrade rat-text to 2.0.0
+* break: remove PagerStyle
 
 # 0.28.1