+++ rat-theme3/changes.md
@@ -1,6 +1,7 @@
 # 2.2.0
 
 * upgrade rat-text to 2.0.0
+* break: remove PagerStyle
 
 # 2.1.1